requirements.txt
flask
flask-cors
python-docx
python-dotenv
google-genai
beautifulsoup4
requests
